requirements.txt
Django==2.1.7
django-bootstrap4==0.0.8
djangorestframework==3.9.2
gunicorn==19.9.0
psycopg2-binary==2.7.7
pytz==2018.9
